% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cvm_files.R
\name{cvm_arqanual}
\alias{cvm_arqanual}
\title{Download de arquivo zip com arquivos XML de vários dias}
\usage{
cvm_arqanual(soapheader)
}
\arguments{
\item{soapheader}{SOAP Header obtido com a função cvm_login}
}
\value{
arquivo zip
}
\description{
Download de arquivo zip com arquivos XML de vários dias
}

