% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cvm_files.R
\name{cvm_arqcomptc}
\alias{cvm_arqcomptc}
\title{Download do arquivo XML de dados diários na data especificada}
\usage{
cvm_arqcomptc(soapheader, date)
}
\arguments{
\item{soapheader}{SOAP Header obtido com a função cvm_login}

\item{date}{data do arquivo}
}
\value{
arquivo XML no diretório atual
}
\description{
Download do arquivo XML de dados diários na data especificada
}

