% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cvm_files.R
\name{cvm_listacomptc}
\alias{cvm_listacomptc}
\title{Lista de arquivos atualmente disponibilizados no sistema}
\usage{
cvm_listacomptc(soapheader, inicdeliverydate = Sys.Date() - 3)
}
\arguments{
\item{soapheader}{SOAP Header obtido com a função cvm_login}

\item{inicdeliverydate}{data inicial de entrega dos documentos}
}
\value{
data.frame com datas dos arquivos disponíveis
}
\description{
Lista de arquivos atualmente disponibilizados no sistema
}

