% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_cvm.R
\name{cvm_login}
\alias{cvm_login}
\title{Efetua o login no sistema da CVM}
\usage{
cvm_login(login, senha)
}
\arguments{
\item{login}{seu login}

\item{senha}{sua senha}
}
\value{
the SOAP header enviada pelo sistema para ser utilizada em outras funções
}
\description{
Efetua o login no sistema da CVM
}

