% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cvm_files.R
\name{read_cvmxmlcad}
\alias{read_cvmxmlcad}
\title{Leitura do arquivo XML de dados cadstrais da CVM}
\usage{
read_cvmxmlcad(file)
}
\arguments{
\item{file}{arquivo XML do servidor da CVM}
}
\value{
data.frame com os dados do arquivo
}
\description{
Leitura do arquivo XML de dados cadstrais da CVM
}

