% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cvm_files.R
\name{read_cvmxmlfile}
\alias{read_cvmxmlfile}
\title{Leitura do arquivo XML da CVM}
\usage{
read_cvmxmlfile(file)
}
\arguments{
\item{file}{arquivo XML do servidor da CVM}
}
\value{
data.frame com os dados do arquivo
}
\description{
Leitura do arquivo XML da CVM
}

